sql: move code for makeOptimizerPlan to new file
Release note: None
